futuragents/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "futuragents"
version = "1.0.0"
description = "AI Multi-Agent Binance Futures Trading System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.1.0",

    # Database & Cache
    "motor>=3.4.0",
    "pymongo>=4.7.0",
    "redis>=5.0.0",

    # Auth & Security
    "PyJWT>=2.8.0",
    "bcrypt>=4.1.0",
    "python-dotenv>=1.0.0",

    # Async & Tasks
    "apscheduler>=3.10.0",
    "aiofiles>=23.2.0",
    "httpx>=0.27.0",

    # Binance & Crypto
    "python-binance>=1.0.19",
    "ccxt>=4.3.0",

    # AI & LLM (Anthropic Claude)
    "anthropic>=0.28.0",
    "langchain-anthropic>=0.1.23",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-experimental>=0.0.62",
    "langgraph>=0.2.0",

    # Technical Analysis
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",
    "pandas-ta>=0.3.14b",
    "stockstats>=0.6.0",

    # Market Data
    "yfinance>=0.2.40",
    "finnhub-python>=2.4.20",
    "requests>=2.32.0",
    "websockets>=12.0",

    # Utils
    "rich>=13.7.0",
    "psutil>=5.9.0",
    "pytz>=2024.1",
    "tqdm>=4.66.0",
    "plotly>=5.22.0",
    "concurrent-log-handler>=0.9.24",
]

[tool.setuptools.packages.find]
include = ["app*"]

[project.scripts]
futuragents = "app.main:start"
